Move exception related files into a folder
