+++ packages/ui/src/features/CourtCaseList/CourtCaseListEntry/CourtCaseListEntry.tsx
@@ -7,8 +7,8 @@ import { DisplayPartialCourtCase } from "types/display/CourtCases"
 import { DisplayFullUser } from "types/display/Users"
 import { deleteQueryParamsByName } from "utils/deleteQueryParam"
 import { canUserUnlockCase } from "utils/formatReasons/canUserUnlockCase"
-import { displayExceptionReasons } from "utils/formatReasons/displayExceptionReasonCell"
-import { canUnlockCase } from "utils/formatReasons/displayExceptionsLockTag"
+import { displayExceptionReasons } from "utils/formatReasons/exceptions/displayExceptionReasonCell"
+import { canUnlockCase } from "utils/formatReasons/exceptions/displayExceptionsLockTag"
 import { formatReasonCodes, ReasonCodes, ReasonCodeTitle } from "utils/formatReasons/reasonCodes"
 import getResolutionStatus from "../../../utils/getResolutionStatus"
 import { CaseDetailsRow } from "./CaseDetailsRow/CaseDetailsRow"
