+++ packages/ui/src/utils/formatReasons/exceptions/displayExceptionsLockTag.ts
@@ -1,5 +1,5 @@
 import type { DisplayFullUser } from "types/display/Users"
-import { canUserUnlockCase } from "./canUserUnlockCase"
+import { canUserUnlockCase } from "../canUserUnlockCase"
 
 export const canUnlockCase = (user: DisplayFullUser, errorLockedByUsername: string | null | undefined): boolean => {
   return !!errorLockedByUsername && canUserUnlockCase(user, errorLockedByUsername)
